refactor(auth): reloacate user queries to `CommandPort` for data consistency
비밀번호 재설정과 같은 작업에서 추후 Read DB와 Command DB 분리 시, user 정보의 정합성을 지키기 위해 Query Port에서 Command Port로 수정함
