+++ services/auth/application/src/main/java/nettee/auth/port/AuthQueryRepositoryPort.java
@@ -4,9 +4,6 @@ import java.util.Optional;
 import nettee.auth.domain.User;
 
 public interface AuthQueryRepositoryPort {
-    Optional<User> findById(String userId);
     Optional<User> findByLoginId(String loginId);
     boolean existsByLoginId(String loginId);
-
-    Optional<User> findByEmail(String email);
 }
