+++ services/auth/application/src/main/java/nettee/auth/service/AuthCommandService.java
@@ -7,7 +7,7 @@ import static nettee.auth.exception.AuthErrorCode.AUTH_ACCOUNT_NOT_FOUND;
 import static nettee.auth.exception.AuthErrorCode.AUTH_OTP_DESERIALIZE_FAILED;
 import static nettee.auth.exception.AuthErrorCode.AUTH_OTP_INVALID;
 import static nettee.auth.exception.AuthErrorCode.AUTH_OTP_SERIALIZE_FAILED;
-import static nettee.auth.exception.AuthErrorCode.AUTH_PASSWORD_NOT_MATCH;
+import static nettee.auth.exception.AuthErrorCode.AUTH_PASSWORD_MISMATCHED;
 import static nettee.auth.exception.AuthErrorCode.AUTH_PASSWORD_RESET_INVALID;
 import static nettee.auth.exception.AuthErrorCode.AUTH_REFRESH_TOKEN_NOT_FOUND;
 
@@ -207,7 +207,7 @@ public class AuthCommandService implements AuthSignUsecase {
     @Override
     public void sendPasswordResetEmail(String email) {
         // 사용자 존재 여부 확인
-        authQueryRepositoryPort.findByEmail(email)
+        authCommandRepositoryPort.findByEmail(email)
                 .orElseThrow(() -> new AuthException(AUTH_ACCOUNT_NOT_FOUND));
 
         // nonce 생성
@@ -224,7 +224,7 @@ public class AuthCommandService implements AuthSignUsecase {
     @Override
     public void resetPassword(String email, String newPassword, String nonce) {
         // 사용자 존재 여부 확인
-        User user = authQueryRepositoryPort.findByEmail(email)
+        User user = authCommandRepositoryPort.findByEmail(email)
                 .orElseThrow(() -> new AuthException(AUTH_ACCOUNT_NOT_FOUND));
 
         // nonce 조회 및 검증
@@ -247,12 +247,12 @@ public class AuthCommandService implements AuthSignUsecase {
     @Override
     public String verifyPassword(String userId, String password) {
         // 비밀번호 조회를 위한 사용자 조회
-        User userEntity = authQueryRepositoryPort.findById(userId)
+        User userEntity = authCommandRepositoryPort.findById(userId)
                 .orElseThrow(() -> new AuthException(AUTH_ACCOUNT_NOT_FOUND));
 
         // 비밀번호 검증
         if (!passwordEncoder.matches(password, userEntity.getEncodedPassword())) {
-            throw new AuthException(AUTH_PASSWORD_NOT_MATCH);
+            throw new AuthException(AUTH_PASSWORD_MISMATCHED);
         }
 
         // nonce 생성 및 redis 저장
@@ -265,7 +265,7 @@ public class AuthCommandService implements AuthSignUsecase {
     @Override
     public void changePassword(String userId, String newPassword, String nonce) {
         // 비밀번호 변경을 위한 사용자 조회
-        User userEntity = authQueryRepositoryPort.findById(userId)
+        User userEntity = authCommandRepositoryPort.findById(userId)
                 .orElseThrow(() -> new AuthException(AUTH_ACCOUNT_NOT_FOUND));
 
         // nonce 조회 및 검증
